backend/app/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic[email]==2.5.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
databases[sqlite]==0.8.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pyjwt==2.8.0
bcrypt==4.1.2